requirements.txt
fastapi==0.111.0
fastapi-cli==0.0.4
pydantic==2.7.4
pydantic_core==2.18.4
openai==1.52.2
python-dotenv==1.0.1
tavily-python==0.5.0
uvicorn==0.30.1